+++ src/FastAccelStepper.h
@@ -182,9 +182,6 @@ class FastAccelStepperEngine {
 #if defined(SUPPORT_RP_PICO)
   uint8_t claimed_pios;
   PIO pio[NUM_PIOS];
-
- public:
-  void pushCommands();
 #endif
 
   friend class FastAccelStepper;